add triggers to `events_stream_ordering` columns in bg job and make this job depend on the `add_stream_ordering` bg job
